Update Node model primary key, fix IPv6 padding bug

+++ database/migrations/2014_10_12_000000_create_users_table.php
@@ -12,9 +12,8 @@ class CreateUsersTable extends Migration {
 	 */
 	public function up()
 	{
-		Schema::create('users', function(Blueprint $table)
-		{
-			$table->increments('id');
+		Schema::create('users', function(Blueprint $table) {
+			$table->bigIncrements('id');
 			$table->string('name');
 			$table->string('email')->unique();
 			$table->string('password', 60);
